+++ Sources/KeyboardKit/SwiftUI/Extensions/Image+Globe.swift
@@ -1,0 +1,20 @@
+//
+//  Image+Globe.swift
+//  KeyboardKit
+//
+//  Created by Daniel Saidi on 2020-03-11.
+//  Copyright © 2020 Daniel Saidi. All rights reserved.
+//
+
+import SwiftUI
+
+@available(iOS 13.0, *)
+public extension Image {
+    
+    /**
+     The globe icon that is used for "next keyboard" buttons.
+     */
+    static var globe: Image {
+        Image(systemName: "globe")
+    }
+}