+++ Sources/KeyboardKit/UIKit/Extensions/UIImage+Globe.swift
@@ -1,0 +1,20 @@
+//
+//  UIImage+Globe.swift
+//  KeyboardKit
+//
+//  Created by Daniel Saidi on 2020-03-11.
+//  Copyright © 2020 Daniel Saidi. All rights reserved.
+//
+
+import UIKit
+
+@available(iOS 13.0, *)
+public extension UIImage {
+    
+    /**
+     The globe icon that is used for "next keyboard" buttons.
+     */
+    static var globe: UIImage? {
+        UIImage(systemName: "globe")
+    }
+}